大四_Senior/DLP/Lab2/requirements.txt
torch==1.10.2
torchvision==0.11.3
torchaudio==0.10.2
numpy>=1.19.2
Pillow>=8.0.0
tqdm>=4.41.0
matplotlib>=3.3.2
albumentations>=0.5.2